feat(tabs): better code structuring, added better types
